x86: refactor boot code
A new src/boot directory is being added, with the boot.h header
file in it, and some definitions from the stage2 code are being
moved to boot.h. This refactoring (no functional changed intended)
will allow code reuse when adding UEFI support (the UEFI bootloader
will run in 64 bit mode).
