+++ platform/pc/boot/def32.h
@@ -12,23 +12,6 @@ typedef u32 word;
 
 #define pointer_from_u64(__a) ((void *)(u32)(__a))
 #define u64_from_pointer(__a) ((u64)(u32)(__a))
-#define physical_from_virtual(__x) u64_from_pointer(__x)
-
-static inline void *tag(void *v, u16 tval)
-{
-    *((u8 *)v-1) = tval;
-    return v;
-}
-
-static inline u16 tagof(void *v)
-{
-    return *((u8 *)v-1);
-}
-
-static inline void *valueof(void *v)
-{
-    return v;
-}
 
 #define DIV(__x, __by, __q, __r)\
  {\
